requirements.txt
selenium
pandas
google-cloud-bigquery
google-auth
google-cloud-storage
pyarrow
scipy
numpy
pytz

